html-source/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.13
langchain-openai>=0.0.5
langchain-qdrant>=0.1.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
openai>=1.12.0
qdrant-client>=1.7.0


